update(api): moved service files into separate folder
