+++ api/src/lib/index.ts
@@ -4,8 +4,9 @@ export {
   RegisterUserSchemaType,
 } from './schemas/registerUserSchema'
 export { LoginUserSchema, LoginUserSchemaType } from './schemas/loginUserSchema'
+export { SendEmailSchema, SendEmailSchemaType } from './schemas/sendEmailSchema'
 export { userService } from './services/db/userService'
 export { subscriptionService } from './services/db/subscriptionService'
 export { AppError } from './errors/app'
-export { EmailService } from './services/emailService'
+export { EmailService } from './services/others/emailService'
 export { createEmailVerificationHtml } from './others/createEmailVerificationHtml'
